event finder files re-structuring
