+++ app/routes/events/all-events/components/all-events-content.component.tsx
@@ -1,25 +1,13 @@
-import { Link, useLoaderData } from 'react-router-dom';
-import { useEffect, useMemo, useState, type RefObject } from 'react';
+import { useLoaderData } from 'react-router-dom';
+import { useEffect, useMemo, useState } from 'react';
 import { liteClient as algoliasearch } from 'algoliasearch/lite';
-import {
-  Configure,
-  InstantSearch,
-  useInstantSearch,
-  usePagination,
-  useRefinementList,
-} from 'react-instantsearch';
+import { Configure, InstantSearch } from 'react-instantsearch';
 
-import { cn } from '~/lib/utils';
 import { createInstantSearchUrlSync } from '~/components/finders/instant-search-url-sync/create-instant-search-url-sync';
-import { useHydratedHitsFallback } from '~/components/finders/use-hydrated-hits-fallback';
-
 import { SectionTitle } from '~/components';
-import { Icon } from '~/primitives/icon/icon';
 import { useScrollToSearchResultsOnLoad } from '~/hooks/use-scroll-to-search-results-on-load';
-import { ResourceCard } from '~/primitives/cards/resource-card';
-import type { ContentItemHit } from '~/routes/search/types';
 
-import type { AllEventsLoaderData, EventFinderFacetItem } from '../loader';
+import type { AllEventsLoaderData } from '../loader';
 import { useEventsAlgoliaRouting } from '../hooks/use-events-algolia-routing.hook';
 import {
   FeaturedEventsFromHits,
@@ -27,167 +15,17 @@ import {
 } from '../partials/featured-events.partial';
 import { parseEventsFinderUrlState } from '../../events-url-state';
 import type { EventsFinderUrlState } from '../../events-url-state';
-import { buildEventSingleUrl } from '../../events-back-url';
-
-import { getEventCardDisplayDate } from './featured-card.component';
-import { EventsFiltersViewport } from './events-filters-viewport.component';
 import {
-  EVENT_FACET_CATEGORIES,
-  EVENT_FACET_LOCATIONS,
   MAIN_EVENTS_GRID_HITS_PER_PAGE,
   MAIN_EVENTS_TYPE_FILTER,
 } from '../all-events.constants';
 
-function getEventHitLocation(
-  hit: ContentItemHit,
-  multipleLabel = 'Multiple Locations',
-) {
-  return hit.eventLocations && hit.eventLocations.length > 1
-    ? multipleLabel
-    : hit.eventLocations?.[0] ||
-        hit.locations?.[0]?.name ||
-        'Christ Fellowship Church';
-}
-
-function MobileEventHitCard({ hit, to }: { hit: ContentItemHit; to: string }) {
-  const displayDate = getEventCardDisplayDate(hit);
-  const location = getEventHitLocation(hit, 'Multiple campuses');
-  const imageUri = hit.coverImage?.sources?.[0]?.uri ?? '';
-
-  return (
-    <Link
-      to={to}
-      className='flex w-full items-center gap-4 overflow-hidden rounded-2xl border border-neutral-lighter bg-white p-2 text-text-primary transition-colors duration-200 hover:border-neutral-light md:hidden'
-      prefetch='intent'
-    >
-      <img
-        src={imageUri}
-        alt={hit.title}
-        className='aspect-square h-24 w-24 shrink-0 rounded-lg object-cover'
-        loading='lazy'
-      />
-
-      <div className='flex min-w-0 flex-1 flex-col gap-2'>
-        <h4 className='line-clamp-2 w-full text-lg font-extrabold leading-tight text-pretty'>
-          {hit.title}
-        </h4>
-
-        <div className='flex min-w-0 items-center gap-2'>
-          <Icon
-            name='calendarAlt'
-            color='currentColor'
-            size={18}
-            className='shrink-0'
-          />
-          <p className='truncate text-sm font-semibold leading-normal'>
-            {displayDate}
-          </p>
-          <Icon
-            name='chevronRight'
-            color='currentColor'
-            size={20}
-            className='ml-auto shrink-0 text-neutral-light'
-          />
-        </div>
-
-        <div className='flex min-w-0 items-center gap-2'>
-          <Icon
-            name='map'
-            color='currentColor'
-            size={18}
-            className='shrink-0'
-          />
-          <p className='truncate text-sm font-semibold leading-normal'>
-            {location}
-          </p>
-        </div>
-      </div>
-    </Link>
-  );
-}
-
-function EventHit({
-  hit,
-  fromEventsUrl,
-}: {
-  hit: ContentItemHit;
-  fromEventsUrl: string;
-}) {
-  const formattedDate = getEventCardDisplayDate(hit);
-
-  const imageUri = hit.coverImage?.sources?.[0]?.uri ?? '';
-  const location = getEventHitLocation(hit);
-
-  // Forward the active finder filters so the event's back button can restore
-  // them (stateless `from` query param — see events-back-url.ts).
-  const eventSingleUrl = buildEventSingleUrl(hit.url, fromEventsUrl);
-
-  return (
-    <>
-      <MobileEventHitCard hit={hit} to={eventSingleUrl} />
-      <div className='hidden h-full md:block'>
-        <ResourceCard
-          resource={{
-            id: hit.objectID,
-            contentChannelId: '78',
-            contentType: 'EVENTS',
-            name: hit.title,
-            summary: hit.summary ?? '',
-            image: imageUri,
-            pathname: eventSingleUrl,
-            startDate: formattedDate,
-            location,
-          }}
-        />
-      </div>
-    </>
-  );
-}
-
-interface PaginationButtonProps {
-  children: React.ReactNode;
-  isDisabled?: boolean;
-  isActive?: boolean;
-  href: string;
-  onClick: () => void;
-  className?: string;
-}
-
-const PaginationButton = ({
-  children,
-  isDisabled = false,
-  isActive = false,
-  href,
-  onClick,
-  className = '',
-}: PaginationButtonProps) => {
-  if (isDisabled) {
-    return (
-      <span
-        className={`${className} inline-flex items-center justify-center border-2 border-neutral-200 text-neutral-200 cursor-not-allowed`}
-      >
-        {children}
-      </span>
-    );
-  }
-
-  return (
-    <a
-      href={href}
-      onClick={(event) => {
-        event.preventDefault();
-        onClick();
-      }}
-      className={`${className} inline-flex items-center justify-center border-2 ${
-        isActive
-          ? 'border-navy bg-navy text-white'
-          : 'border-navy text-navy hover:bg-navy hover:text-white'
-      } transition-colors duration-200 cursor-pointer`}
-    >
-      {children}
-    </a>
-  );
-};
+import { AllEventsInstantFilters } from './all-events-instant-filters.component';
+import {
+  AllEventsInstantResults,
+  AllEventsResultsLayout,
+} from './all-events-results.component';
+import { EventsFiltersViewport } from './events-filters-viewport.component';
 
 export function AllEventsContent() {
   const {
@@ -360,197 +198,5 @@ export function AllEventsContent() {
         )}
       </div>
     </>
-  );
-}
-
-function mapRefinementItemsToFacets(
-  items: ReturnType<typeof useRefinementList>['items'],
-): EventFinderFacetItem[] {
-  // Existing events filter components are presentational and expect loader-like
-  // facet objects. Map InstantSearch refinement items into that same shape so
-  // desktop and mobile UI can stay unchanged.
-  return items.map((item) => ({
-    value: item.value,
-    label: item.label,
-    count: item.count,
-  }));
-}
-
-function AllEventsInstantFilters({
-  eventsMobilePinEndRef,
-  buildAllEventsInstantSearchUiState,
-}: Pick<
-  Parameters<typeof EventsFiltersViewport>[0],
-  'eventsMobilePinEndRef'
-> & {
-  buildAllEventsInstantSearchUiState: (
-    urlState: EventsFinderUrlState,
-  ) => Record<string, Record<string, unknown>>;
-}) {
-  const { indexUiState, setIndexUiState } = useInstantSearch();
-  const categoryItems = useRefinementList({
-    attribute: EVENT_FACET_CATEGORIES,
-    limit: 50,
-  });
-  const locationItems = useRefinementList({
-    attribute: EVENT_FACET_LOCATIONS,
-    limit: 50,
-  });
-
-  // Events filter components still work with URL-shaped state because that is
-  // also what the active-chip and clear-all UI understand. Here the source of
-  // truth is InstantSearch's index state, converted into the same shape.
-  const urlState = {
-    query:
-      typeof indexUiState.query === 'string' &&
-      indexUiState.query.trim().length > 0
-        ? indexUiState.query
-        : undefined,
-    refinementList:
-      (indexUiState.refinementList as Record<string, string[]>) ?? undefined,
-    page:
-      typeof indexUiState.page === 'number' && indexUiState.page > 0
-        ? indexUiState.page
-        : undefined,
-  } satisfies EventsFinderUrlState;
-
-  const applyUrlState = (next: EventsFinderUrlState) => {
-    // Convert the events UI's URL-shaped update back into InstantSearch state.
-    // The parent `onStateChange` then writes the canonical URL, so filter clicks
-    // remain client-side without invoking the route loader.
-    const nextState = buildAllEventsInstantSearchUiState(next);
-    const nextUiState = Object.values(nextState)[0];
-    setIndexUiState((nextUiState ?? {}) as Record<string, unknown>);
-  };
-
-  return (
-    <EventsFiltersViewport
-      onClearAllToUrl={() => setIndexUiState({})}
-      eventsMobilePinEndRef={eventsMobilePinEndRef}
-      urlState={urlState}
-      applyUrlState={applyUrlState}
-      categoryFacets={mapRefinementItemsToFacets(categoryItems.items)}
-      locationFacets={mapRefinementItemsToFacets(locationItems.items)}
-    />
-  );
-}
-
-function AllEventsInstantResults({
-  initialEventHits,
-  fromEventsUrl,
-}: {
-  initialEventHits: ContentItemHit[];
-  fromEventsUrl: string;
-}) {
-  const { hits, isLoading } = useHydratedHitsFallback<ContentItemHit>({
-    initialHits: initialEventHits,
-  });
-  const { currentRefinement, nbPages, isFirstPage, isLastPage, refine } =
-    usePagination();
-
-  // The index response is already ordered by Algolia's custom ranking
-  // (eventStartDate), so hits render in index order.
-  const eventHits = hits;
-
-  return (
-    <AllEventsResultsLayout
-      eventHits={eventHits}
-      eventsNbPages={nbPages}
-      eventsPage={currentRefinement}
-      isFirstPage={isFirstPage}
-      isLastPage={isLastPage}
-      isLoading={isLoading}
-      fromEventsUrl={fromEventsUrl}
-      goToPage={(nextPage) => {
-        // Let InstantSearch own pagination after hydration; the route URL is
-        // synchronized by `onStateChange` above.
-        refine(Math.max(0, nextPage));
-        window.requestAnimationFrame(() => {
-          const scrollTarget = document.querySelector('.pagination-scroll-to');
-          scrollTarget?.scrollIntoView({ behavior: 'smooth', block: 'start' });
-        });
-      }}
-    />
-  );
-}
-
-function AllEventsResultsLayout({
-  eventHits,
-  eventsNbPages,
-  eventsPage,
-  isFirstPage,
-  isLastPage,
-  isLoading,
-  fromEventsUrl,
-  goToPage,
-  eventsMobilePinEndRef,
-}: {
-  eventHits: ContentItemHit[];
-  eventsNbPages: number;
-  eventsPage: number;
-  isFirstPage: boolean;
-  isLastPage: boolean;
-  isLoading: boolean;
-  fromEventsUrl: string;
-  goToPage: (nextPage: number) => void;
-  eventsMobilePinEndRef?: RefObject<HTMLDivElement | null>;
-}) {
-  return (
-    <div className='content-padding pt-16 md:pt-0'>
-      <div className='mx-auto min-h-88 w-full max-w-screen-content md:min-h-112'>
-        {eventHits.length === 0 && !isLoading ? null : (
-          <ul
-            className={cn(
-              'grid w-full list-none grid-cols-1 justify-items-center gap-4 p-0 md:grid-cols-2 md:gap-10 lg:grid-cols-3',
-              isLoading && 'opacity-60 pointer-events-none',
-            )}
-            role='list'
-          >
-            {eventHits.map((hit) => (
-              <li key={hit.objectID} className='w-full'>
-                <EventHit hit={hit} fromEventsUrl={fromEventsUrl} />
-              </li>
-            ))}
-          </ul>
-        )}
-
-        {eventsNbPages > 1 ? (
-          <div className='flex items-center justify-start gap-4 mt-16'>
-            <PaginationButton
-              isDisabled={isFirstPage}
-              onClick={() => goToPage(eventsPage - 1)}
-              href='#'
-              className='w-12 h-12'
-            >
-              <Icon name='chevronLeft' size={24} />
-            </PaginationButton>
-
-            <PaginationButton
-              isActive
-              onClick={() => {}}
-              href='#'
-              className='w-12 h-12 bg-navy text-white'
-            >
-              {eventsPage + 1}
-            </PaginationButton>
-
-            <PaginationButton
-              isDisabled={isLastPage}
-              onClick={() => goToPage(eventsPage + 1)}
-              href='#'
-              className='w-12 h-12'
-            >
-              <Icon name='chevronRight' size={24} />
-            </PaginationButton>
-          </div>
-        ) : null}
-
-        <div
-          ref={eventsMobilePinEndRef}
-          className='pointer-events-none h-0 w-full shrink-0'
-          aria-hidden
-        />
-      </div>
-    </div>
   );
 }
